feat: updated content and improved styles
