--- layouts/blogpost.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class='min-h-screen bg-gray-50 dark:bg-gray-900'>
-    <UContainer>
-      <article class='prose dark:prose-invert max-w-none'>
-        <slot />
-      </article>
-    </UContainer>
-  </div>
-</template>
