+++ package.json
@@ -10,8 +10,8 @@
   "devDependencies": {
     "@iconify-json/heroicons": "^1.2.2",
     "@iconify-json/simple-icons": "^1.2.26",
-    "@nuxt/content": "^2.13.4",
-    "@nuxt/eslint": "^0.7.6",
+    "@nuxt/content": "^3.2.2",
+    "@nuxt/eslint": "^1.1.0",
     "@nuxt/fonts": "^0.10.3",
     "@nuxt/image": "^1.9.0",
     "@nuxt/ui": "^2.21.0",
@@ -24,15 +24,12 @@
     "@tresjs/core": "^4.3.3",
     "@tresjs/nuxt": "^3.0.8",
     "@types/node": "^22.13.5",
-    "@types/three": "^0.173.0",
+    "@types/three": "^0.174.0",
     "@vercel/analytics": "^1.5.0",
     "gsap": "^3.12.7",
     "nuxt": "^3.15.4",
     "nuxt-disqus": "^1.1.1",
     "nuxt-llms": "^0.1.0",
-    "three": "^0.173.0"
-  },
-  "dependencies": {
-    "nuxt-content-assets": "^1.4.4"
+    "three": "^0.174.0"
   }
 }
